services/events_service/requirements.txt
alembic==1.15.2
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.0
elastic-transport==8.17.1
elasticsearch==9.0.2
Flask==3.1.1
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
kafka-python==2.2.15
Mako==1.3.10
MarkupSafe==3.0.2
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
redis==6.2.0
requests==2.32.3
six==1.17.0
SQLAlchemy==2.0.41
typing_extensions==4.13.2
urllib3==2.4.0
Werkzeug==3.1.3
